// lib/types/pageTypes.ts
export type ServiceCardProps = {
  title: string
  description: string
}

export type TimelineItemProps = {
  date: string
  title: string
  description: string
}

export type PortfolioItemProps = {
  title: string
  description: string
  imageUrl: string
  link?: string
}

export type TestimonialProps = {
  name: string
  company: string
  testimonial: string
  rating: number
} 